requirements.txt
pandas==1.3.2
torchio==0.18.47
torch==1.9.0
PyYAML==5.4.1
protobuf==3.17.3
medpy==0.4.0